requirements.txt
flask
flask_cors
pychromecast
pillow
opencv-python
requests
numpy

